requirements.txt
streamlit
psycopg2-binary
torch
torchvision
pillow
opencv-python
pytesseract